refactored Relabel;ByHashing
