+++ src/main/java/org/cicirello/sequences/distance/internal/RelabelByHashing.java
@@ -218,390 +218,4 @@ public final class RelabelByHashing implements KendallTauRelabeler {
     }
     return current;
   }
-
-  static class BaseHT {
-    protected final int mask;
-    protected final int minSize;
-
-    BaseHT(int maxSize, int minSize) {
-      final int MAX_SIZE = maxSize;
-      if (minSize > MAX_SIZE) {
-        minSize = MAX_SIZE;
-        mask = minSize - 1;
-      } else {
-        minSize = minSize - 1;
-        minSize = minSize | (minSize >> 1);
-        minSize = minSize | (minSize >> 2);
-        minSize = minSize | (minSize >> 4);
-        minSize = minSize | (minSize >> 8);
-        minSize = minSize | (minSize >> 16);
-        mask = minSize;
-        minSize++;
-      }
-      this.minSize = minSize;
-    }
-  }
-
-  private static final class IntHT extends BaseHT {
-
-    private final Node[] table;
-
-    IntHT(int min) {
-      super(0x40000000, min);
-      table = new Node[minSize];
-    }
-
-    int populate(int[] s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length; i++) {
-        if (!containsKey(s1[i])) {
-          current++;
-          put(s1[i], current);
-        }
-      }
-      return current;
-    }
-
-    int index(int key) {
-      return (key ^ (key >>> 16)) & mask;
-    }
-
-    boolean containsKey(int key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return true;
-      }
-      return false;
-    }
-
-    int get(int key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return current.value;
-      }
-      // NOTE: our internal usage never puts a negative as a value
-      return -1;
-    }
-
-    void put(int key, int value) {
-      // warning: assumes key is not already in hash table (only used internally so ok).
-      int i = index(key);
-      table[i] = new Node(key, value, table[i]);
-    }
-
-    static final class Node {
-      int key;
-      int value;
-      Node next;
-
-      Node(int key, int value, Node next) {
-        this.key = key;
-        this.value = value;
-        this.next = next;
-      }
-    }
-  }
-
-  private static final class LongHT extends BaseHT {
-
-    private final Node[] table;
-
-    LongHT(int min) {
-      super(0x40000000, min);
-      table = new Node[minSize];
-    }
-
-    int populate(long[] s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length; i++) {
-        if (!containsKey(s1[i])) {
-          current++;
-          put(s1[i], current);
-        }
-      }
-      return current;
-    }
-
-    int index(long key) {
-      int x = (int) (key ^ (key >>> 32));
-      return (x ^ (x >>> 16)) & mask;
-    }
-
-    boolean containsKey(long key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return true;
-      }
-      return false;
-    }
-
-    int get(long key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return current.value;
-      }
-      // NOTE: our internal usage never puts a negative as a value
-      return -1;
-    }
-
-    void put(long key, int value) {
-      // warning: assumes key is not already in hash table (only used internally so ok).
-      int i = index(key);
-      table[i] = new Node(key, value, table[i]);
-    }
-
-    static final class Node {
-      long key;
-      int value;
-      Node next;
-
-      Node(long key, int value, Node next) {
-        this.key = key;
-        this.value = value;
-        this.next = next;
-      }
-    }
-  }
-
-  private static final class ShortHT extends BaseHT {
-
-    private final Node[] table;
-
-    ShortHT(int min) {
-      super(0x10000, min);
-      table = new Node[minSize];
-    }
-
-    int populate(short[] s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length; i++) {
-        if (!containsKey(s1[i])) {
-          current++;
-          put(s1[i], current);
-        }
-      }
-      return current;
-    }
-
-    int index(short key) {
-      return key & mask;
-    }
-
-    boolean containsKey(short key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return true;
-      }
-      return false;
-    }
-
-    int get(short key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return current.value;
-      }
-      // NOTE: our internal usage never puts a negative as a value
-      return -1;
-    }
-
-    void put(short key, int value) {
-      // warning: assumes key is not already in hash table (only used internally so ok).
-      int i = index(key);
-      table[i] = new Node(key, value, table[i]);
-    }
-
-    static final class Node {
-      short key;
-      int value;
-      Node next;
-
-      Node(short key, int value, Node next) {
-        this.key = key;
-        this.value = value;
-        this.next = next;
-      }
-    }
-  }
-
-  private static final class CharHT extends BaseHT {
-
-    private final Node[] table;
-
-    CharHT(int min) {
-      super(0x10000, min);
-      table = new Node[minSize];
-    }
-
-    int populate(char[] s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length; i++) {
-        if (!containsKey(s1[i])) {
-          current++;
-          put(s1[i], current);
-        }
-      }
-      return current;
-    }
-
-    int populate(String s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length(); i++) {
-        if (!containsKey(s1.charAt(i))) {
-          current++;
-          put(s1.charAt(i), current);
-        }
-      }
-      return current;
-    }
-
-    int index(char key) {
-      return key & mask;
-    }
-
-    boolean containsKey(char key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return true;
-      }
-      return false;
-    }
-
-    int get(char key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return current.value;
-      }
-      // NOTE: our internal usage never puts a negative as a value
-      return -1;
-    }
-
-    void put(char key, int value) {
-      // warning: assumes key is not already in hash table (only used internally so ok).
-      int i = index(key);
-      table[i] = new Node(key, value, table[i]);
-    }
-
-    static final class Node {
-      char key;
-      int value;
-      Node next;
-
-      Node(char key, int value, Node next) {
-        this.key = key;
-        this.value = value;
-        this.next = next;
-      }
-    }
-  }
-
-  private static final class DoubleHT extends BaseHT {
-
-    private final Node[] table;
-
-    DoubleHT(int min) {
-      super(0x40000000, min);
-      table = new Node[minSize];
-    }
-
-    int populate(double[] s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length; i++) {
-        if (!containsKey(s1[i])) {
-          current++;
-          put(s1[i], current);
-        }
-      }
-      return current;
-    }
-
-    int index(double key) {
-      long x = Double.doubleToLongBits(key);
-      int y = (int) (x ^ (x >>> 32));
-      return (y ^ (y >>> 16)) & mask;
-    }
-
-    boolean containsKey(double key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return true;
-      }
-      return false;
-    }
-
-    int get(double key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return current.value;
-      }
-      // NOTE: our internal usage never puts a negative as a value
-      return -1;
-    }
-
-    void put(double key, int value) {
-      // warning: assumes key is not already in hash table (only used internally so ok).
-      int i = index(key);
-      table[i] = new Node(key, value, table[i]);
-    }
-
-    static final class Node {
-      double key;
-      int value;
-      Node next;
-
-      Node(double key, int value, Node next) {
-        this.key = key;
-        this.value = value;
-        this.next = next;
-      }
-    }
-  }
-
-  private static final class FloatHT extends BaseHT {
-
-    private final Node[] table;
-
-    FloatHT(int min) {
-      super(0x40000000, min);
-      table = new Node[minSize];
-    }
-
-    int populate(float[] s1) {
-      int current = -1;
-      for (int i = 0; i < s1.length; i++) {
-        if (!containsKey(s1[i])) {
-          current++;
-          put(s1[i], current);
-        }
-      }
-      return current;
-    }
-
-    int index(float key) {
-      int x = Float.floatToIntBits(key);
-      return (x ^ (x >>> 16)) & mask;
-    }
-
-    boolean containsKey(float key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return true;
-      }
-      return false;
-    }
-
-    int get(float key) {
-      for (Node current = table[index(key)]; current != null; current = current.next) {
-        if (current.key == key) return current.value;
-      }
-      // NOTE: our internal usage never puts a negative as a value
-      return -1;
-    }
-
-    void put(float key, int value) {
-      // warning: assumes key is not already in hash table (only used internally so ok).
-      int i = index(key);
-      table[i] = new Node(key, value, table[i]);
-    }
-
-    static final class Node {
-      float key;
-      int value;
-      Node next;
-
-      Node(float key, int value, Node next) {
-        this.key = key;
-        this.value = value;
-        this.next = next;
-      }
-    }
-  }
 }
