+++ src/test/java/org/cicirello/sequences/distance/internal/KendallTauSequenceDistanceInternalTests.java
@@ -28,33 +28,152 @@ import org.junit.jupiter.api.*;
 public class KendallTauSequenceDistanceInternalTests {
 
   @Test
-  public void testKendallTauSequenceDistance_HashTableBaseClass() {
-    class TestHT extends RelabelByHashing.BaseHT {
-      TestHT(int min) {
-        super(32, min);
-      }
+  public void testKendallTauSequenceDistance_HashTableInt() {
+    final int max = 0x40000000;
+    int[] powerOfTwoCases = {1, 2, 4, 8, 16, 32};
+    for (int n : powerOfTwoCases) {
+      IntHT ht = new IntHT(n);
+      assertEquals(n, ht.size());
+      assertEquals(n - 1, ht.mask());
     }
-    for (int n = 1; n <= 32; n *= 2) {
-      TestHT ht = new TestHT(n);
-      assertEquals(n, ht.minSize);
-      assertEquals(n - 1, ht.mask);
-    }
-    TestHT ht = new TestHT(3);
-    assertEquals(4, ht.minSize);
+    IntHT ht = new IntHT(3);
+    assertEquals(4, ht.size());
+    assertEquals(max, ht.adjustSize(max + 1));
+    assertEquals(max, ht.adjustSize(max));
     for (int n = 5; n < 8; n++) {
-      ht = new TestHT(n);
-      assertEquals(8, ht.minSize);
-      assertEquals(7, ht.mask);
+      ht = new IntHT(n);
+      assertEquals(8, ht.size());
+      assertEquals(7, ht.mask());
     }
     for (int n = 9; n < 16; n++) {
-      ht = new TestHT(n);
-      assertEquals(16, ht.minSize);
-      assertEquals(15, ht.mask);
+      ht = new IntHT(n);
+      assertEquals(16, ht.size());
+      assertEquals(15, ht.mask());
     }
-    for (int n = 17; n <= 64; n++) {
-      ht = new TestHT(n);
-      assertEquals(32, ht.minSize);
-      assertEquals(31, ht.mask);
+  }
+
+  @Test
+  public void testKendallTauSequenceDistance_HashTableLong() {
+    final int max = 0x40000000;
+    int[] powerOfTwoCases = {1, 2, 4, 8, 16, 32};
+    for (int n : powerOfTwoCases) {
+      LongHT ht = new LongHT(n);
+      assertEquals(n, ht.size());
+      assertEquals(n - 1, ht.mask());
+    }
+    LongHT ht = new LongHT(3);
+    assertEquals(4, ht.size());
+    assertEquals(max, ht.adjustSize(max + 1));
+    assertEquals(max, ht.adjustSize(max));
+    for (int n = 5; n < 8; n++) {
+      ht = new LongHT(n);
+      assertEquals(8, ht.size());
+      assertEquals(7, ht.mask());
+    }
+    for (int n = 9; n < 16; n++) {
+      ht = new LongHT(n);
+      assertEquals(16, ht.size());
+      assertEquals(15, ht.mask());
+    }
+  }
+
+  @Test
+  public void testKendallTauSequenceDistance_HashTableShort() {
+    final int max = 0x40000000;
+    int[] powerOfTwoCases = {1, 2, 4, 8, 16, 32};
+    for (int n : powerOfTwoCases) {
+      ShortHT ht = new ShortHT(n);
+      assertEquals(n, ht.size());
+      assertEquals(n - 1, ht.mask());
+    }
+    ShortHT ht = new ShortHT(3);
+    assertEquals(4, ht.size());
+    assertEquals(max, ht.adjustSize(max + 1));
+    assertEquals(max, ht.adjustSize(max));
+    for (int n = 5; n < 8; n++) {
+      ht = new ShortHT(n);
+      assertEquals(8, ht.size());
+      assertEquals(7, ht.mask());
+    }
+    for (int n = 9; n < 16; n++) {
+      ht = new ShortHT(n);
+      assertEquals(16, ht.size());
+      assertEquals(15, ht.mask());
+    }
+  }
+
+  @Test
+  public void testKendallTauSequenceDistance_HashTableChar() {
+    final int max = 0x40000000;
+    int[] powerOfTwoCases = {1, 2, 4, 8, 16, 32};
+    for (int n : powerOfTwoCases) {
+      CharHT ht = new CharHT(n);
+      assertEquals(n, ht.size());
+      assertEquals(n - 1, ht.mask());
+    }
+    CharHT ht = new CharHT(3);
+    assertEquals(4, ht.size());
+    assertEquals(max, ht.adjustSize(max + 1));
+    assertEquals(max, ht.adjustSize(max));
+    for (int n = 5; n < 8; n++) {
+      ht = new CharHT(n);
+      assertEquals(8, ht.size());
+      assertEquals(7, ht.mask());
+    }
+    for (int n = 9; n < 16; n++) {
+      ht = new CharHT(n);
+      assertEquals(16, ht.size());
+      assertEquals(15, ht.mask());
+    }
+  }
+
+  @Test
+  public void testKendallTauSequenceDistance_HashTableDouble() {
+    final int max = 0x40000000;
+    int[] powerOfTwoCases = {1, 2, 4, 8, 16, 32};
+    for (int n : powerOfTwoCases) {
+      DoubleHT ht = new DoubleHT(n);
+      assertEquals(n, ht.size());
+      assertEquals(n - 1, ht.mask());
+    }
+    DoubleHT ht = new DoubleHT(3);
+    assertEquals(4, ht.size());
+    assertEquals(max, ht.adjustSize(max + 1));
+    assertEquals(max, ht.adjustSize(max));
+    for (int n = 5; n < 8; n++) {
+      ht = new DoubleHT(n);
+      assertEquals(8, ht.size());
+      assertEquals(7, ht.mask());
+    }
+    for (int n = 9; n < 16; n++) {
+      ht = new DoubleHT(n);
+      assertEquals(16, ht.size());
+      assertEquals(15, ht.mask());
+    }
+  }
+
+  @Test
+  public void testKendallTauSequenceDistance_HashTableFloat() {
+    final int max = 0x40000000;
+    int[] powerOfTwoCases = {1, 2, 4, 8, 16, 32};
+    for (int n : powerOfTwoCases) {
+      FloatHT ht = new FloatHT(n);
+      assertEquals(n, ht.size());
+      assertEquals(n - 1, ht.mask());
+    }
+    FloatHT ht = new FloatHT(3);
+    assertEquals(4, ht.size());
+    assertEquals(max, ht.adjustSize(max + 1));
+    assertEquals(max, ht.adjustSize(max));
+    for (int n = 5; n < 8; n++) {
+      ht = new FloatHT(n);
+      assertEquals(8, ht.size());
+      assertEquals(7, ht.mask());
+    }
+    for (int n = 9; n < 16; n++) {
+      ht = new FloatHT(n);
+      assertEquals(16, ht.size());
+      assertEquals(15, ht.mask());
     }
   }
 }
